+++ package.json
@@ -7,6 +7,7 @@
     "type-check": "tsc"
   },
   "dependencies": {
+    "axios": "^0.27.2",
     "next": "latest",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
